changed some code
For the index page: I changed the position of players and cards, so when users adjust their browser size, the positions will have the same position (keep the ratio depending on the browser size). However, I failed to adjust the size of the background image. The background image's position is wrong for now.

For the log-in page: I changed the size of input boxes for username and password. Even though users change the size of their browser, the box size will not change. Also, I adjust the background image's size. The background image will no more repeat, and the size will automatically adjust depending on the size of the browser.